Add callback for stop etcd connection

+++ registry/etcdv3/register.go
@@ -2,28 +2,24 @@ package etcdv3
 
 import (
 	"strings"
+	"time"
 
 	"github.com/luoyancn/dubhe/logging"
 	etcdconf "github.com/luoyancn/dubhe/registry/etcdv3/config"
 
 	etcd "github.com/coreos/etcd/clientv3"
+	uuid "github.com/satori/go.uuid"
 	"golang.org/x/net/context"
 	"google.golang.org/grpc"
 )
 
 var deregister = make(chan struct{})
 
-type Option struct {
-	RegistryDir string
-	ServiceName string
-	NodeID      string
-	NData       string
-}
-
-func Register(config etcd.Config, opt Option) error {
-	key := strings.Join(
-		[]string{opt.RegistryDir, opt.ServiceName, opt.NodeID}, "/")
-	logging.LOG.Infof("Register service with key :%s\n", key)
+func generate_etcd_config() etcd.Config {
+	config := etcd.Config{
+		Endpoints:   etcdconf.ETCD_ENDPOINTS,
+		DialTimeout: etcdconf.ETCD_CONNECTION_TIMEOUT,
+	}
 
 	// 1.no dial timeout means New doesn't block and connection attempt
 	// happens in the background.
@@ -33,26 +29,40 @@ func Register(config etcd.Config, opt Option) error {
 	// lientv3.New() won't return error when no endpoint is available.
 	// https://github.com/coreos/etcd/issues/9829
 	// https://github.com/coreos/etcd/issues/9877
-	if config.DialTimeout > 0 {
+	if etcdconf.ETCD_CONNECTION_TIMEOUT > 0 {
 		logging.LOG.Warningf("With DialTimeout options, we must use blocking\n")
-		logging.LOG.Warningf(" call to Ensure the error would be returned\n")
-		logging.LOG.Warningf(" Visit the follow links to get more details:\n")
-		logging.LOG.Warningf(" https://github.com/coreos/etcd/issues/9829\n")
-		logging.LOG.Warningf(" https://github.com/coreos/etcd/issues/9877\n")
+		logging.LOG.Warningf("call to Ensure the error would be returned\n")
+		logging.LOG.Warningf("Visit the follow links to get more details:\n")
+		logging.LOG.Warningf("https://github.com/coreos/etcd/issues/9829\n")
+		logging.LOG.Warningf("https://github.com/coreos/etcd/issues/9877\n")
 		config.DialOptions = append(config.DialOptions, grpc.WithBlock())
 	}
+	return config
+}
+
+func Register(ndata string) error {
+	config := generate_etcd_config()
+	nodeid, _ := uuid.NewV4()
+	key := strings.Join(
+		[]string{etcdconf.ETCD_REGISTER_DIR,
+			etcdconf.ETCD_SERVICE_NAME, nodeid.String()}, "/")
+	logging.LOG.Infof("Register service with key :%s\n", key)
+
+	var client *etcd.Client
+	go func() {
+		<-deregister
+		if nil != client {
+			logging.LOG.Infof("Deleting %s from etcd...\n", key)
+			client.Delete(context.Background(), key)
+		}
+		deregister <- struct{}{}
+	}()
+
 	client, err := etcd.New(config)
 	if err != nil {
 		logging.LOG.Fatalf("Cannot connect to endpoins :%v in %d seconds\n",
-			config.Endpoints, config.DialTimeout)
+			config.Endpoints, config.DialTimeout/time.Second)
 	}
-
-	go func() {
-		<-deregister
-		logging.LOG.Infof("Deleting %s from etcd...\n", key)
-		client.Delete(context.Background(), key)
-		deregister <- struct{}{}
-	}()
 
 	grant_ctx, grant_cancle := context.WithTimeout(
 		context.Background(), etcdconf.ETCD_CONNECTION_TIMEOUT)
@@ -67,16 +77,16 @@ func Register(config etcd.Config, opt Option) error {
 		context.Background(), etcdconf.ETCD_CONNECTION_TIMEOUT)
 	defer put_cancle()
 	if _, err = client.Put(put_ctx, key,
-		opt.NData, etcd.WithLease(resp.ID)); err != nil {
+		ndata, etcd.WithLease(resp.ID)); err != nil {
 		logging.LOG.Errorf(
-			"Failed Put the key %s value %sinto etcd:%v\n", key, opt.NData, err)
+			"Failed Put the key %s value %sinto etcd:%v\n", key, ndata, err)
 		return err
 	}
 
 	keep_ctx, keep_cancle := context.WithTimeout(
 		context.Background(), etcdconf.ETCD_CONNECTION_TIMEOUT)
 	defer keep_cancle()
-	if _, err := client.KeepAlive(keep_ctx, resp.ID); err != nil {
+	if _, err = client.KeepAlive(keep_ctx, resp.ID); err != nil {
 		logging.LOG.Errorf(
 			"Failed refresh the key %s exsit in etcd:%v\n", key, err)
 		return err
